Refactor out AbiEncoder
