+++ .github/autolabeler.yml
@@ -12,6 +12,7 @@ types: ['packages/types']
 instant: ['packages/instant']
 abi-gen-templates: ['packages/abi-gen-templates']
 abi-gen: ['packages/abi-gen']
+abi-encoder: ['packages/abi-encoder']
 website: ['packages/website']
 sol-cov: ['packages/sol-cov']
 utils: ['packages/utils']
