+++ packages/abi-encoder/src/abi_encoder/abstract_data_types/interfaces.ts
