+++ packages/abi-encoder/src/abi_encoder/abstract_data_types/types/blob.ts
