+++ packages/abi-encoder/src/abi_encoder/calldata/iterator.ts
