+++ packages/abi-encoder/src/abi_encoder/evm_data_types/int.ts
@@ -1,58 +1,59 @@
+import { BigNumber } from '@0x/utils';
 import { DataItem, SolidityTypes } from 'ethereum-types';
 import * as _ from 'lodash';
 
-import { BigNumber } from '../../configured_bignumber';
 import { DataTypeFactory } from '../abstract_data_types/interfaces';
 import { AbstractBlobDataType } from '../abstract_data_types/types/blob';
 import { RawCalldata } from '../calldata/raw_calldata';
 import { constants } from '../utils/constants';
 import * as EncoderMath from '../utils/math';
 
-export class UIntDataType extends AbstractBlobDataType {
+export class IntDataType extends AbstractBlobDataType {
     private static readonly _MATCHER = RegExp(
-        '^uint(8|16|24|32|40|48|56|64|72|88|96|104|112|120|128|136|144|152|160|168|176|184|192|200|208|216|224|232|240|248|256){0,1}$',
+        '^int(8|16|24|32|40|48|56|64|72|88|96|104|112|120|128|136|144|152|160|168|176|184|192|200|208|216|224|232|240|248|256){0,1}$',
     );
     private static readonly _SIZE_KNOWN_AT_COMPILE_TIME: boolean = true;
     private static readonly _MAX_WIDTH: number = 256;
-    private static readonly _DEFAULT_WIDTH: number = UIntDataType._MAX_WIDTH;
-    private static readonly _MIN_VALUE = new BigNumber(0);
+    private static readonly _DEFAULT_WIDTH: number = IntDataType._MAX_WIDTH;
     private readonly _width: number;
+    private readonly _minValue: BigNumber;
     private readonly _maxValue: BigNumber;
 
     public static matchType(type: string): boolean {
-        return UIntDataType._MATCHER.test(type);
+        return IntDataType._MATCHER.test(type);
     }
 
     private static _decodeWidthFromType(type: string): number {
-        const matches = UIntDataType._MATCHER.exec(type);
+        const matches = IntDataType._MATCHER.exec(type);
         const width =
             !_.isNull(matches) && matches.length === 2 && !_.isUndefined(matches[1])
                 ? parseInt(matches[1], constants.DEC_BASE)
-                : UIntDataType._DEFAULT_WIDTH;
+                : IntDataType._DEFAULT_WIDTH;
         return width;
     }
 
     public constructor(dataItem: DataItem, dataTypeFactory: DataTypeFactory) {
-        super(dataItem, dataTypeFactory, UIntDataType._SIZE_KNOWN_AT_COMPILE_TIME);
-        if (!UIntDataType.matchType(dataItem.type)) {
-            throw new Error(`Tried to instantiate UInt with bad input: ${dataItem}`);
+        super(dataItem, dataTypeFactory, IntDataType._SIZE_KNOWN_AT_COMPILE_TIME);
+        if (!IntDataType.matchType(dataItem.type)) {
+            throw new Error(`Tried to instantiate Int with bad input: ${dataItem}`);
         }
-        this._width = UIntDataType._decodeWidthFromType(dataItem.type);
-        this._maxValue = new BigNumber(2).toPower(this._width).sub(1);
+        this._width = IntDataType._decodeWidthFromType(dataItem.type);
+        this._minValue = new BigNumber(2).toPower(this._width - 1).times(-1);
+        this._maxValue = new BigNumber(2).toPower(this._width - 1).sub(1);
     }
 
     public encodeValue(value: BigNumber | string | number): Buffer {
-        const encodedValue = EncoderMath.safeEncodeNumericValue(value, UIntDataType._MIN_VALUE, this._maxValue);
+        const encodedValue = EncoderMath.safeEncodeNumericValue(value, this._minValue, this._maxValue);
         return encodedValue;
     }
 
     public decodeValue(calldata: RawCalldata): BigNumber {
         const valueBuf = calldata.popWord();
-        const value = EncoderMath.safeDecodeNumericValue(valueBuf, UIntDataType._MIN_VALUE, this._maxValue);
+        const value = EncoderMath.safeDecodeNumericValue(valueBuf, this._minValue, this._maxValue);
         return value;
     }
 
     public getSignature(): string {
-        return `${SolidityTypes.Uint}${this._width}`;
+        return `${SolidityTypes.Int}${this._width}`;
     }
 }
