+++ packages/abi-encoder/src/abi_encoder/evm_data_types/string.ts
