+++ packages/abi-encoder/src/abi_encoder/index.ts
@@ -1,4 +1,3 @@
-export { EncodingRules, DecodingRules } from './utils/rules';
 export {
     Address,
     Array,
@@ -12,3 +11,4 @@ export {
     Tuple,
     UInt,
 } from './evm_data_type_factory';
+export { DecodingRules, EncodingRules } from './utils/rules';
