+++ packages/abi-encoder/src/abi_encoder/utils/constants.ts
