+++ packages/abi-encoder/src/abi_encoder/utils/math.ts
