+++ packages/abi-encoder/src/index.ts
@@ -1,0 +1,2 @@
+import * as AbiEncoder from './abi_encoder';
+export { AbiEncoder };
