+++ packages/abi-encoder/tsconfig.json
@@ -1,0 +1,8 @@
+{
+    "extends": "../../tsconfig",
+    "compilerOptions": {
+        "outDir": "lib",
+        "rootDir": "."
+    },
+    "include": ["src/**/*", "test/**/*"]
+}
