+++ packages/abi-encoder/tslint.json
@@ -1,0 +1,3 @@
+{
+    "extends": ["@0x/tslint-config"]
+}
